requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
requests==2.32.0
pyngrok==7.2.0
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.27.0
livekit-agents[google]>=1.3.0
livekit-api>=1.0.0
